+++ MiniWebServer.MiniApp/Content/ByteArrayContent.cs
@@ -23,7 +23,7 @@ namespace MiniWebServer.MiniApp.Content
 
         public override HttpHeaders Headers => headers;
 
-        public override async Task<int> WriteToAsync(IBufferWriter<byte> writer, CancellationToken cancellationToken)
+        public override async Task<long> WriteToAsync(IBufferWriter<byte> writer, CancellationToken cancellationToken)
         {
             writer.Write(content);
