+++ MiniWebServer.MiniApp/MiniContent.cs
@@ -14,7 +14,7 @@ namespace MiniWebServer.MiniApp
     public abstract class MiniContent: IHttpContent
     {
         public abstract Abstractions.Http.HttpHeaders Headers { get; }
-        public abstract Task<int> WriteToAsync(IBufferWriter<byte> writer, CancellationToken cancellationToken);
+        public abstract Task<long> WriteToAsync(IBufferWriter<byte> writer, CancellationToken cancellationToken);
         public abstract long ContentLength { get; }
     }
 }
